Moved functions into helper namespace
